1. Finally that stupid (not-quite-so-)"invisible" test plot for getting the plotregeion could be completely removed. It served only to implement square pixels, which is since so,e IDL version supported by the /ISOTROPIC keyword.
Only thing that changed is the units in which LEGMARGIN is specified.

2.
Removed temporary resetting of !P.Multi to 1. Don't know why this is
in there. Now !P.Multi works as expected again.
KNOWN BUG: plotting a legend overwrites a portion of the window. This
is a known bug in TvSclLegend I will get back to when I have the time.